+++ map_simple.cub
@@ -9,10 +9,11 @@ C 200,200,0
 111111111111111111111111111111111111111111111
 100000000000000000000000000000000000000000001
 100000000000000000000000000000000000000000001
-100000000000000000000000000000000000000000001
-100000000000000000000000000000000000000000001
-100000000000000000N00000000000000000000000001
-100000000000000000000000000000000000000000001
-100000000000000000000000000000000000000000001
+100000000011111111111111000000000000000000001
+100000000001100000000001000000000000000000001
+100000000111000000N00001000000000000000000001
+100000000001000000000111000000000000000000001
+100000000001000000000111000000000000000000001
+100000000001111111111110000000000000000000001
 100000000000000000000000000000000000000000001
 111111111111111111111111111111111111111111111
