+++ src/main.c
@@ -11,7 +11,6 @@
 /* ************************************************************************** */
 
 #include "game.h"
-#include <math.h>
 
 // #define PI 3.141592
 
@@ -21,7 +20,6 @@ double deg_to_rad(double deg)
 }
 
 ///TODO
-	// regulate map to width - pad with 1s
 	// deg to radians ?
 
 uint32_t	get(uint8_t *texture_pixels)
@@ -142,87 +140,6 @@ void mark_pnt(t_gameInfo	*gi, int x, int y, uint32_t color)
 
 }
 
-double get_horiz_dist(t_gameInfo	*gi, double angle)
-{
-
-	if (angle == 0 || angle == 180)
-		return 0;
-
-	int U_D = 1;
-	if (angle < 180)
-		U_D = 0;
-	int L_R = 0;
-	if (angle < 90 || angle > 270)
-		L_R = 1;
-	printf("L_R: %i, UD = %i\n", L_R, U_D);
-
-	//PLAYER XY
-	int P_x = gi->player->x + (PLAYER_SIZE / 2);
-	int P_y = gi->player->y + (PLAYER_SIZE / 2);
-
-
-	//FIRST INTERSECTION XY
-	int B_x;
-	int B_y;
-
-
-	B_y = (P_y / TEXTURE_SIZE) * TEXTURE_SIZE;
-	B_x = B_y / tan(deg_to_rad(angle));
-	if (angle == 90 || angle == 270)
-		B_x = P_x;
-
-
-	double dis_diff = TEXTURE_SIZE / cos(deg_to_rad(angle));
-			printf("DIS diff: %f\n", dis_diff);
-
-
-	//how much to add to the x pos in each loop
-	// double x_diff = TEXTURE_SIZE * tan (deg_to_rad(angle));
-	double x_diff = dis_diff * cos (deg_to_rad(angle));
-	/* 	if (angle == 45 || angle == 315)
-			y_diff *= -1; */
-
-
-
-	double	dis;
-	if (U_D == 0 && L_R == 0)
-		dis = (TEXTURE_SIZE - (P_y % TEXTURE_SIZE)) / sin(deg_to_rad(180 - angle));
-	if (U_D == 1 && L_R == 0)
-		dis = (TEXTURE_SIZE - (P_y % TEXTURE_SIZE)) / sin(deg_to_rad(angle - 180));
-	if (U_D == 0 && L_R == 1)
-		dis = (TEXTURE_SIZE - (P_y % TEXTURE_SIZE)) / sin(deg_to_rad(angle));
-	if (U_D == 1 && L_R == 1)
-		dis = (TEXTURE_SIZE - (P_y % TEXTURE_SIZE)) / sin(deg_to_rad(360 - angle));
-
-/* 	if(angle == 225 || angle == 135)
-		dis += dis_diff; */
-
-
-	int	y_diff = TEXTURE_SIZE;
-	if (U_D == 0) //up
-		y_diff *= -1;
-
-
-
-	printf("x diff: %f, tan: (%f) y diff: %i\n", x_diff,tan (deg_to_rad(angle)), y_diff);
-	printf("P(XY): %i, %i\n\n", P_x, P_y);
-	printf("B(XY): %i, %i\n", B_x, B_y);
-
-
-	while (B_x < WIDTH && B_x > 0 && B_y > 0 && B_y < HEIGHT)
-	{
-		mark_pnt(gi, B_x, B_y, 0xFF00FFFF);
-
-		if (coors_in_map(gi, B_x, B_y) == '1')
-			return (fabs(dis));
-		B_x += x_diff;
-		B_y += y_diff;
-		dis += dis_diff;
-		// printf("B(XY): %i, %i\n", B_x, B_y);
-	}
-	return (fabs(dis));
-}
-
 
 void	print_screen(t_gameInfo *game_info)
 {
@@ -236,27 +153,51 @@ void	print_screen(t_gameInfo *game_info)
 	}
 	mlx_delete_image(game_info->mlx, game_info->screen_image);
 	game_info->screen_image = img;
+/*
+	double vert_dis = get_vert_dist(game_info,(double)game_info->player->orientation);
 
+	double horiz_dis = get_horiz_dist(game_info,(double)game_info->player->orientation);
 
+	double dis = vert_dis;
+	if (!vert_dis)
+		dis = horiz_dis;
+	else if (!horiz_dis)
+		dis = vert_dis;
+	else if (horiz_dis < vert_dis)
+		dis = horiz_dis;
+	else if (vert_dis < horiz_dis)
+		dis = vert_dis;
 
-	double dis;
-	dis = get_horiz_dist(game_info,(double)game_info->player->orientation);
 	for(int i = 0; i < (int) dis; i++)
 	{
 		draw_dot(game_info,(double)game_info->player->orientation, i);
-	}
-/* 	double arr[] = {0, 45, 90, 135, 180, 225, 270, 315};
+	} */
 
-	 for(int a = 0; a < 8; a++)
+	double vert_dis, horiz_dis, dis;
+	double arr[] = {0, 45, 90, 135, 180, 225, 270, 315};
+
+	for(int a = 0; a < 8; a++)
 	{
-		// dis = get_vert_dist(game_info, arr[a]);
-		dis = get_horiz_dist(game_info, arr[a]);
+		vert_dis = get_vert_dist(game_info,arr[a]);
+
+		horiz_dis = get_horiz_dist(game_info,arr[a]);
+
+		dis = vert_dis;
+		if (!vert_dis)
+			dis = horiz_dis;
+		else if (!horiz_dis)
+			dis = vert_dis;
+		else if (horiz_dis < vert_dis)
+			dis = horiz_dis;
+		else if (vert_dis < horiz_dis)
+			dis = vert_dis;
+
 		for(int i = 0; i < (int) dis; i++)
 		{
 			draw_dot(game_info,arr[a], i);
 		}
 	}
- */
+
 	mlx_image_to_window(game_info->mlx, game_info->screen_image, 0, 0);
 }
 
@@ -277,98 +218,3 @@ int	main(int argc, char *argv[])
 	return (EXIT_SUCCESS);
 }
 
-
-
-
-/* double get_vert_dist(t_gameInfo	*gi, double angle)
-{
-
-	if (angle == 90 || angle == 270)
-		return 0;
-
-	int U_D = 1;
-	if (angle < 180)
-		U_D = 0;
-
-	int L_R = 0;
-	if (angle < 90 || angle > 270)
-		L_R = 1;
-
-	printf("L_R: %i, UD = %i\n", L_R, U_D);
-
-	//PLAYER XY
-	int P_x = gi->player->x + (PLAYER_SIZE / 2);
-	int P_y = gi->player->y + (PLAYER_SIZE / 2);
-	//FIRST INTERSECTION XY
-	int B_x;
-	int B_y;
-
-	if (L_R == 1) // RIGHT
-		B_x = ((int)(P_x / TEXTURE_SIZE) + 1) * TEXTURE_SIZE;
-	else
-		B_x = ((int)(P_x / TEXTURE_SIZE)) * TEXTURE_SIZE;
-
-	if (L_R == 1)
-	{
-		if (U_D == 0)
-			B_y = P_y + (P_x-B_x) * tan(deg_to_rad(angle)); // top right
-		else
-			B_y = P_y - (P_x-B_x) * tan(deg_to_rad(360 - angle)); // bottom right
-	}
-
-	if (L_R == 0)
-	{
-		if (U_D == 0)
-			B_y = P_y - (P_x-B_x) * tan(deg_to_rad(180 - angle)); // top left
-		else
-			B_y = P_y + (P_x-B_x) * tan(deg_to_rad(angle - 180)); // bottom left
-	}
-	//how much to add to the y pos in each loop
-	double y_diff = TEXTURE_SIZE * tan (deg_to_rad(angle));
-	if (angle == 45 || angle == 315)
-		y_diff *= -1;
-
-
-	//first value is the hypoteneuse of player to the nearest check point
-	double dis_diff = TEXTURE_SIZE / cos(deg_to_rad(angle));
-			printf("DIS diff: %f\n", dis_diff);
-		//first value is the hypoteneuse of player to the nearest check point
-
-	double	dis;
-	if (U_D == 0 && L_R == 0)
-		dis = (TEXTURE_SIZE - (P_x % TEXTURE_SIZE)) / cos(deg_to_rad(180 - angle));
-	if (U_D == 1 && L_R == 0)
-		dis = (TEXTURE_SIZE - (P_x % TEXTURE_SIZE)) / cos(deg_to_rad(angle - 180));
-	if (U_D == 0 && L_R == 1)
-		dis = (TEXTURE_SIZE - (P_x % TEXTURE_SIZE)) / cos(deg_to_rad(angle));
-	if (U_D == 1 && L_R == 1)
-		dis = (TEXTURE_SIZE - (P_x % TEXTURE_SIZE)) / cos(deg_to_rad(360 - angle));
-
-	if(angle == 225 || angle == 135)
-		dis += dis_diff;
-
-
-	int	x_diff = TEXTURE_SIZE;
-	if (L_R == 0) //left
-		x_diff *= -1;
-	printf("x diff: %i, tan: (%f) y diff: %f\n", x_diff,tan (deg_to_rad(angle)), y_diff);
-	printf("P(XY): %i, %i\n\n", P_x, P_y);
-	printf("B(XY): %i, %i\n", B_x, B_y);
-
-
-	while (B_x < WIDTH && B_x > 0 && B_y > 0 && B_y < HEIGHT)
-	{
-
-
-		mark_pnt(gi, B_x, B_y, 0xFF00FFFF);
-
-		if (coors_in_map(gi, B_x, B_y) == '1')
-			return (fabs(dis));
-		B_x += x_diff;
-		B_y += y_diff;
-		dis += dis_diff;
-		// printf("B(XY): %i, %i\n", B_x, B_y);
-	}
-	return (fabs(dis));
-}
- */
