+++ src/player_movement.c
@@ -19,6 +19,9 @@ char	coors_in_map(t_gameInfo *gi, int x, int y)
 {
 	if (x < 0 || y < 0 || x > WIDTH || y > HEIGHT)
 		return (0);
+
+/* 	printf("xy: %i(%i), %i (%i)\n", x, x % TEXTURE_SIZE, y, y % TEXTURE_SIZE);
+	printf("BLOC: %i, %i\n", x / TEXTURE_SIZE, y / TEXTURE_SIZE); */
 	return (gi->map_info->map
 		[y / TEXTURE_SIZE]
 		[x / TEXTURE_SIZE]);
